requirements.txt
numpy
ModernGL
Pillow
wxPython